chore: update diagnosis status

+++ src/main/java/com/upc/proyecto1/channel/model/aggregate/DiagnosisStatuOutput.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class DiagnosisStatuOutput {
-  private Integer countDiagnosis;
-  private Long countSampleDiagnosis;
+  private Long countDiagnosis;
+  private Integer countSampleDiagnosis;
 
 }
